Port code from images.py to test_arima.py because AppVeyor was having difficulty with relative imports.
